Symbol and footprint for ESP-WROOM-32 module
As per the datasheet
[here](https://espressif.com/sites/default/files/documentation/esp_wroom_32_datasheet_en.pdf).

The ESP32 is quite flexible in that it can remap most alternate
functions onto any of the pins. So I've decided to only label the
significant pins on the symbol.

The datasheet is suprisingly good but it inconsistently refers to the
exposed pad as pin 1 on the diagram (fig 1) and pin 39 in the table
(table 2). I've gone with 39.