New STM32F1 parts
This adds symbols for the 64- and 100-pin variants of the
STM32F105/107. (For some reason I accidentally thought I wanted
the 100-pin variant but it's a behemoth.)

Pin naming is imported from the ISIS files and unmodified except for
changing NRST to ~RST and reordering.

I assume the footprints will be as standard but I haven't
comprehensively checked them...